Move getPreferredWildcard to /src/lib
