+++ src/versionmanager.ts
@@ -12,6 +12,7 @@ import * as versionUtil from './version-util'
 import packageManagers from './package-managers'
 import { supportedVersionTargets } from './constants'
 import { FilterPattern, GetVersion, IgnoredUpgrade, Index, Maybe, Options, PackageManager, PackageFile, Version, VersionDeclaration } from './types'
+import getPreferredWildcard from './lib/getPreferredWildcard'
 
 interface MappedDependencies {
   current: VersionDeclaration,
@@ -67,41 +68,6 @@ function isUpgradeable(current: VersionDeclaration, latest: Version) {
     isValidLatest &&
     !isSatisfied(latestNormalized, range.operator === '<' ? current : version) &&
     !semver.ltr(latestNormalized, version)
-}
-
-/**
- *
- * @param dependencies A dependencies collection
- * @returns Returns whether the user prefers ^, ~, .*, or .x
- * (simply counts the greatest number of occurrences) or `null` if
- * given no dependencies.
- */
-function getPreferredWildcard(dependencies: Index<string>) {
-
-  // if there are no dependencies, return null.
-  if (Object.keys(dependencies).length === 0) {
-    return null
-  }
-
-  // group the dependencies by wildcard
-  const groups = _.groupBy(Object.values(dependencies), dep =>
-    versionUtil.WILDCARDS.find((wildcard: string) =>
-      dep && dep.includes(wildcard)
-    )
-  )
-
-  delete groups.undefined // eslint-disable-line fp/no-delete
-
-  // convert to an array of objects that can be sorted
-  const arrOfGroups = cint.toArray<string[], { wildcard: string, instances: string[] }>(groups, (wildcard, instances) => ({
-    wildcard,
-    instances
-  }))
-
-  // reverse sort the groups so that the wildcard with the most appearances is at the head, then return it.
-  const sorted = _.sortBy(arrOfGroups, wildcardObject => -wildcardObject.instances.length)
-
-  return sorted.length > 0 ? sorted[0].wildcard : null
 }
 
 /**
@@ -627,7 +593,6 @@ module.exports = {
   getOwnerPerDependency,
 
   // exposed for testing
-  getPreferredWildcard,
   isUpgradeable,
   queryVersions,
   upgradeDependencies,
